move the commands to a seperate cpp file
